+++ app/streammetadata.h
@@ -11,6 +11,7 @@ public:
   bool isEmpty() const;
 
   void insert(const QString& key, const QString &value);
+  void setStatusNowPlaying(const QString &raw);
   void clear();
 
   quint16 bitrate() const;
@@ -22,9 +23,12 @@ public:
   QString format() const;
 
 private:
+  QString nowPlaying() const;
+
   QMap<QString, QString> _data;
 
   QString title_explicit;
+  QString _status_now_playing;
 };
 
 #endif // STREAMMETADATA_H